style: misc updates
